chore: create format.go in java package
