+++ internal/librarian/java/format.go
@@ -1,0 +1,67 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package java
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+
+	"github.com/googleapis/librarian/internal/command"
+	"github.com/googleapis/librarian/internal/config"
+)
+
+// Format formats a Java client library using google-java-format.
+func Format(ctx context.Context, library *config.Library) error {
+	files, err := collectJavaFiles(library.Output)
+	if err != nil {
+		return fmt.Errorf("failed to find java files for formatting: %w", err)
+	}
+	if len(files) == 0 {
+		return nil
+	}
+
+	if _, err := exec.LookPath("google-java-format"); err != nil {
+		return fmt.Errorf("google-java-format not found in PATH: %w", err)
+	}
+
+	args := append([]string{"--replace"}, files...)
+	if err := command.Run(ctx, "google-java-format", args...); err != nil {
+		return fmt.Errorf("failed to format files: %w", err)
+	}
+	return nil
+}
+
+func collectJavaFiles(root string) ([]string, error) {
+	var files []string
+	err := filepath.WalkDir(root, func(path string, d os.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() || filepath.Ext(path) != ".java" {
+			return nil
+		}
+		// exclude samples/snippets/generated
+		if strings.Contains(path, filepath.Join("samples", "snippets", "generated")) {
+			return nil
+		}
+		files = append(files, path)
+		return nil
+	})
+	return files, err
+}
